chore(styles): rename file to CamelCase and user 'name + Action' for api call
