--- client/src/api/init.localStorage.js
@@ -1,6 +1,0 @@
-const LOCAL_STORAGE_KEY = {
-  isLogged: "@atlana/logged",
-  cartLocal: "@atlana/cart",
-};
-
-export default LOCAL_STORAGE_KEY;
