--- client/src/components/Input/InputErrorMessage.js
@@ -1,11 +1,0 @@
-const ERROR_MESSAGES = {
-  nameInvalid: "Cannot be empty! Only letters allowed.",
-  emailInvalid: "Only gmail is allowed!",
-  emailExisted: "Email is already used.",
-  lengthError: "Length allow 8 to 32",
-  emptyError: "Empty not allow",
-  numberError: "Enter number only",
-  passwordWeak: "Password weak or too long!",
-};
-
-export default ERROR_MESSAGES;
