+++ client/src/pages/Purchase/Desktop/OrderDashboard.jsx
@@ -4,7 +4,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { format, addDays } from "date-fns";
 
 import { formatCapitalize } from "../../../utils/formatData";
-import { getOrderListReq } from "../../../features/order/orderRequest";
+import { getOrderListReq } from "../../../features/order/orderAction";
 import {
   convertStateToStateCode,
   formatMoney,
