+++ client/src/utils/constVariable.js
@@ -36,10 +36,27 @@ const EMAIL_TYPE = {
   unlockAccount: "unlock",
 };
 
+const INPUT_ERROR_MESSAGE = {
+  nameInvalid: "Cannot be empty! Only letters allowed.",
+  emailInvalid: "Only gmail is allowed!",
+  emailExisted: "Email is already used.",
+  lengthError: "Length allow 8 to 32",
+  emptyError: "Empty not allow",
+  numberError: "Enter number only",
+  passwordWeak: "Password weak or too long!",
+};
+
+const LOCAL_STORAGE_KEY = {
+  isLogged: "@atlana/logged",
+  cartLocal: "@atlana/cart",
+};
+
 export {
   ORDER_STATE_MAP,
   ORDER_STATE_CODE,
   ORDER_STATE_ORIGINAL,
+  INPUT_ERROR_MESSAGE,
+  LOCAL_STORAGE_KEY,
   COMMON_PATH,
   EMAIL_TYPE,
 };
